moved merge into compile expressions and now expressions work in settings :)
