+++ pax-example/.pax/cartridge/src/lib.rs
@@ -32,18 +32,8 @@ pub fn instantiate_expression_table<R: 'static + RenderContext>() -> HashMap<usi
     let mut vtable: HashMap<usize, Box<dyn Fn(ExpressionContext<R>) -> TypesCoproduct>> = HashMap::new();
 
     
-    //Color::rgb(0.1,0.3,0)
-    vtable.insert(0, Box::new(|ec: ExpressionContext<R>| -> TypesCoproduct {
-        
-
-        #[allow(unused_parens)]
-        TypesCoproduct::__pax_stdCOCOtypesCOCOColor(
-            Color::rgb((Numeric::from(0.1)),(Numeric::from(0.3)),(Numeric::from(0)),)
-        )
-    }));
-    
     //Transform2D::align(50%,50%)*Transform2D::anchor(50%,50%)*Transform2D::rotate(rotation)
-    vtable.insert(1, Box::new(|ec: ExpressionContext<R>| -> TypesCoproduct {
+    vtable.insert(0, Box::new(|ec: ExpressionContext<R>| -> TypesCoproduct {
         
             let rotation = {
                 let properties = if let Some(sf) = (*ec.stack_frame).borrow().nth_ancestor(0) {
@@ -53,19 +43,31 @@ pub fn instantiate_expression_table<R: 'static + RenderContext>() -> HashMap<usi
                 }.borrow().deref().get_properties();
                 let properties = &*(*properties).borrow();
 
-                if let PropertiesCoproduct::HelloRGB(p) = properties {
-                    
-                    *p.rotation.get()
-                    
-                } else {
-                    unreachable!("1")
-                }
+                
+                    if let PropertiesCoproduct::HelloRGB(p) = properties {
+                        
+                            Numeric::from(p.rotation.get())
+                        
+                    } else {
+                        unreachable!("0")
+                    }
+                
             };
         
 
         #[allow(unused_parens)]
         TypesCoproduct::Transform2D(
-            ((Transform2D::align((Size::Percent(50.into())),(Size::Percent(50.into())),)*Transform2D::anchor((Size::Percent(50.into())),(Size::Percent(50.into())),))*Transform2D::rotate(((rotation).into()),))
+            ((Transform2D::align((Size::Percent(50.into())),(Size::Percent(50.into())),)*(Transform2D::anchor((Size::Percent(50.into())),(Size::Percent(50.into())),)).into())*(Transform2D::rotate((rotation),)).into())
+        )
+    }));
+    
+    //Color::rgb(0.4,0.5,0)
+    vtable.insert(1, Box::new(|ec: ExpressionContext<R>| -> TypesCoproduct {
+        
+
+        #[allow(unused_parens)]
+        TypesCoproduct::__pax_stdCOCOtypesCOCOColor(
+            Color::rgb((Numeric::from(0.4)),(Numeric::from(0.5)),(Numeric::from(0)),)
         )
     }));
     
@@ -81,23 +83,10 @@ pub fn instantiate_root_component<R: 'static + RenderContext>(instance_registry:
         properties: PropertiesCoproduct::HelloRGB( HelloRGB::default() ),
         handler_registry:  Some(Rc::new(RefCell::new(
                                                      HandlerRegistry {
-                                                         click_handlers: vec![
-                                                                    |stack_frame, args|{
-                                                                        let properties = ((*stack_frame).borrow().get_properties());
-                                                                        let properties = &mut *properties.as_ref().borrow_mut();
-                                                                        let properties = if let PropertiesCoproduct::HelloRGB(p) = properties {p} else {unreachable!()};
-                                                                        HelloRGB::handle_global_click(properties,args);
-                                                                    },
-                                                                ],
+                                                         click_handlers: vec![],
                                                          will_render_handlers: vec![],
-                                                         scroll_handlers: vec![
-                                                                     |stack_frame, args|{
-                                                                         let properties = ((*stack_frame).borrow().get_properties());
-                                                                         let properties = &mut *properties.as_ref().borrow_mut();
-                                                                         let properties = if let PropertiesCoproduct::HelloRGB(p) = properties {p} else {unreachable!()};
-                                                                         HelloRGB::handle_global_scroll(properties,args);
-                                                                     },
-                                                                 ],
+                                                         did_mount_handlers: vec![],
+                                                         scroll_handlers: vec![],
                                                      }
                                                  ))),
         instance_registry: Rc::clone(&instance_registry),
@@ -112,7 +101,7 @@ pax_std_primitives::ellipse::EllipseInstance::instantiate(
         
             stroke: Box::new( PropertyLiteral::new(Default::default()) ),
         
-            fill: Box::new( PropertyExpression::new(0) ),
+            fill: Box::new( PropertyExpression::new(1) ),
         
     }),
     handler_registry:  Some(Rc::new(RefCell::new(
@@ -124,6 +113,7 @@ pax_std_primitives::ellipse::EllipseInstance::instantiate(
                                                                     HelloRGB::handle_click(properties,args);
                                                                 },],
                                                      will_render_handlers: vec![],
+                                                     did_mount_handlers: vec![],
                                                      scroll_handlers: vec![|stack_frame, args|{
                                                                      let properties = (*stack_frame).borrow().get_properties();
                                                                      let properties = &mut *properties.as_ref().borrow_mut();
@@ -133,9 +123,9 @@ pax_std_primitives::ellipse::EllipseInstance::instantiate(
                                                  }
                                              ))),
     instance_registry: Rc::clone(&instance_registry),
-    transform: Rc::new(RefCell::new(PropertyExpression::new(1))),
+    transform: Rc::new(RefCell::new(PropertyExpression::new(0))),
     size: Some(Rc::new(RefCell::new(
-        [Box::new(PropertyLiteral::new(Size::Percent(1 .into()))),Box::new(PropertyLiteral::new(Size::Percent(100.into())))]
+        [Box::new(PropertyLiteral::new(Size::Percent(50.into()))),Box::new(PropertyLiteral::new(Size::Percent(50.into())))]
     ))),
     children: Some(Rc::new(RefCell::new(vec![
         
@@ -143,15 +133,16 @@ pax_std_primitives::ellipse::EllipseInstance::instantiate(
     component_template: None,
     scroller_args: None,
     slot_index: None,
-    repeat_source_expression: None,
+    repeat_source_expression_vec: None,
+    repeat_source_expression_range: None,
     conditional_boolean_expression: None,
     compute_properties_fn: None,
 })
-
 ]))),
         scroller_args: None,
         slot_index: None,
-        repeat_source_expression: None,
+        repeat_source_expression_vec: None,
+        repeat_source_expression_range: None,
         conditional_boolean_expression: None,
         compute_properties_fn: Some(Box::new(|properties, rtc|{
             let properties = &mut *properties.as_ref().borrow_mut();
